refactor: move to_kebab into name.rs (name-format home)
CLAUDE.md (wado-compiler): "Use utilities in name.rs to handle name
mangling … Other components must not know the details of name formats."
The Wado→CM kebab-case converter lived in component_model.rs and was
imported across the codebase; move it to name.rs, the designated home for
naming logic. Pure relocation; call sites just re-point their import.
