+++ wado-compiler/src/name.rs
@@ -60,6 +60,41 @@ pub fn namespace_member_alias(namespace: &str, member: &str) -> String {
     format!("{namespace}{NAMESPACE_MEMBER_SEP}{member}")
 }
 
+/// Convert a Wado identifier (`snake_case` / `PascalCase` / `camelCase`) to
+/// Component Model kebab-case (`my-api`, `http-server`, `error-code`).
+pub(crate) fn to_kebab(name: &str) -> String {
+    let chars: Vec<char> = name.chars().collect();
+    let mut out = String::with_capacity(name.len() + 4);
+    let mut prev_lower_or_digit = false;
+    for (i, &ch) in chars.iter().enumerate() {
+        if ch == '_' {
+            if !out.ends_with('-') && !out.is_empty() {
+                out.push('-');
+            }
+            prev_lower_or_digit = false;
+            continue;
+        }
+        if ch.is_ascii_uppercase() {
+            // Break before an uppercase letter that starts a new word: either
+            // after a lowercase/digit (`myApi` -> `my-api`), or at the end of
+            // an acronym run when the next char is lowercase
+            // (`HTTPServer` -> `http-server`).
+            let acronym_boundary = chars.get(i + 1).is_some_and(char::is_ascii_lowercase)
+                && i > 0
+                && chars[i - 1].is_ascii_uppercase();
+            if (prev_lower_or_digit || acronym_boundary) && !out.is_empty() {
+                out.push('-');
+            }
+            out.push(ch.to_ascii_lowercase());
+            prev_lower_or_digit = false;
+        } else {
+            out.push(ch);
+            prev_lower_or_digit = ch.is_ascii_lowercase() || ch.is_ascii_digit();
+        }
+    }
+    out
+}
+
 /// Prefix the compiler stamps onto every synthesised closure-functor
 /// struct (`__Closure_0`, `__Closure_1`, …). Like
 /// [`CLOSURE_CALL_METHOD`], this is purely a compiler-internal
